requirements.txt
google-generativeai
Pillow
pandas
matplotlib
seaborn
langchain
langchain-google-genai
langgraph
langchain-core